Move RunMode code, Use DSL overloads
